Extremely early days.
Create 1000 transforms. Every frame, set their position to a value.

~ 165 fps according to editor fps counter.

I'm guessing that blitting out to transforms is quite expensive and probably the last thing you want to do. Literally. It might be better to blit to a particle effect which doesn't have as much overhead as transforms? Not sure.

Also been trying to make Assembly Definitions for the first time. It's slightly more work to expose necessary libraries to the the Assembly space you've set up, but only slightly, and this should all scale pretty nice.

Used example work from
* https://github.com/stella3d/job-system-cookbook/
* https://www.youtube.com/watch?v=KJZoSV-JX5I&feature=youtu.be&ab_channel=ThousandAnt

* https://www.youtube.com/watch?v=vLrGdx-Ob3g&ab_channel=InfallibleCode - Namespaces (prerequisite for Assemby Definitions)
* https://www.youtube.com/watch?v=HYqOSkHI674&ab_channel=InfallibleCode - Assembly Definitions