Added Edge search
Edge search goes up to an arbitrary limit because I couldn't find a great way to
predetermine the size of lists of blobs. Edge search is still O^N (blob vs blob).
Needed a [NativeDisableParallelForRestriction] to be able to access parts of the array that were outside the current ChunkSize forparallel job. Thanks nick!

Added burst compile attribute to all jobs. Wow. Like 5 times faster. Obvs I have some optimisation to do for nearest neighbour stuff (probably not in the distant future because working out the game feel is going to rely heavily on the frame rate of the goo). Thanks again nick!

Added a little bit of debug viz for the goo blobs to show how many edges it had, and built a new unlit shader using the shader node system, which was so so easy to use, especially having written HLSL most of my career.

Co-Authored-By: Nick Ludlam <nick@recoil.org>