Just some Debug lines  to show unrequited pairings
When we search for "closest N" points we can very often have a point that doesn't include its neighbour, which has included it. So you can have "conga lines" of points in some cases. This causes a large part of the destabilisation. Hopefully a generous amount of neighbours, plus a voronoi search, can make sure we only care about voronoi nearest neighbours.